Add I piece and fix piece offsets
The I piece was a huge oversight. It moves very awkwardly though, will
likely need to find a better starting position or change some physics
(hand wave).

The piece offsets have been bugging me for awhile. Most noticeably,
after adding a spawn point game object and spawning the pieces at its
location, the cube piece was offset by a unit of space. I had no idea
why, but it was infuriating. It turns out it was because the empty game
object housing the individual cubes was offset slightly from the origin.
This is also why the spawn point location needed to be 1.5 Y, instead of
0. I don't know how to move the parent object without messing with the
children objects, so I destroyed the prefabs, separated the children
from their parents, reset the parent position, added the children
back, and recreated the prefabs. This seems to worked well.