Smooth camera!
Turns out that the jerkiness of the camera was not the camera at all, but
the pieces moving through the air in a rough fashion. Updating the piece
rigid bodies to interpolate transform changes caused the camera to be
stable while the piece flies through the air. Exciting!

Also exposes a yOffset property on the camera.