Squash the cube piece
The cube has stood out for awhile as the only piece to use more than 4
blocks (with 8). I didn't think much about it when I first created the
cube, but after messing around with the game for a bit, the fact that it
doesn't follow the 4 block "rule" has been bugging me. This trims the
cube down to 4 pieces, so it's the same width as the other pieces. It's
kind of weird in three dimensions, but it now follows the other pieces and
more closely resembles the original tetrominos.