Increase control in the air
With the physicsy movement of the pieces, it's pretty easy to have too
much or too little velocity when jumping off the runway. Having more
movement speed while airborne can help the player recover, but it feels
like a bandage. I'd like to solve the underlying problem, but because it's
just "physics" I don't know have any ideas of how to do that.