Better height calculation
The pieces consist off 4 blocks composed together within an empty game
object. Calculating the max Y on the empty game object will always be the
center of geometry, based on the composed children (at least I think
that's what it was doing). Instead, we can use the mesh renderer of each
of the individual blocks to determine the highest point.

This increases the number of iterations by a factor of four. I'm somewhat
concerned about performance, but I don't know if I should worry about it
or not. I'll keep an eye out.