Start scene with a flyby
This took way more effort than I expected it to. The Unity animator is
pretty obtuse and does not seem to work well at all for very simple cases,
which I think this flyby could be described as. I'm sure there's a better
way to do what I've done, but at this point I'm tired of dealing with the
less than desirable animation docs and the often unhelpful community
answers to relevant questions. This seems to work good enough.